+++ sootup.jimple.frontend/src/main/java/sootup/jimple/frontend/LazyJimpleMethodSource.java
@@ -25,6 +25,7 @@ package sootup.jimple.frontend;
 import java.nio.file.Path;
 import java.util.*;
 import org.jspecify.annotations.NonNull;
+import sootup.core.IdentifierFactory;
 import sootup.core.frontend.BodySource;
 import sootup.core.frontend.ResolveException;
 import sootup.core.graph.MutableBlockControlFlowGraph;
@@ -33,6 +34,9 @@ import sootup.core.jimple.basic.SimpleStmtPositionInfo;
 import sootup.core.jimple.basic.StmtPositionInfo;
 import sootup.core.jimple.common.*;
 import sootup.core.jimple.common.expr.Expr;
+import sootup.core.jimple.basic.*;
+import sootup.core.jimple.common.constant.*;
+import sootup.core.jimple.common.expr.*;
 import sootup.core.jimple.common.stmt.BranchingStmt;
 import sootup.core.jimple.common.stmt.Stmt;
 import sootup.core.model.*;
@@ -61,6 +65,7 @@ public class LazyJimpleMethodSource implements BodySource {
   private final @NonNull List<ClassType> exceptions;
   @NonNull private final Position methodPosition;
   private Body _bodySource;
+  private final IdentifierFactory identifierFactory;
 
   public LazyJimpleMethodSource(
       @NonNull MethodSignature methodSignature,
@@ -83,6 +88,7 @@ public class LazyJimpleMethodSource implements BodySource {
     this.methodModifiers = methodModifiers;
     this.exceptions = exceptions;
     this.methodPosition = methodPosition;
+    this.identifierFactory = view.getIdentifierFactory();
     this._bodySource = null;
   }
 
@@ -117,10 +123,6 @@ public class LazyJimpleMethodSource implements BodySource {
     private final HashMap<BranchingStmt, List<String>> unresolvedBranches = new HashMap<>();
     private final HashMap<String, Stmt> labeledStmts = new HashMap<>();
     private HashMap<String, Local> locals = new HashMap<>();
-
-    public Local getLocal(@NonNull String name) {
-      return locals.computeIfAbsent(name, (ignored) -> new Local(name, UnknownType.getInstance()));
-    }
 
     @NonNull
     public Body parseMethodBody() {
@@ -181,7 +183,10 @@ public class LazyJimpleMethodSource implements BodySource {
       }
 
       // statements
-      StmtVisitor stmtVisitor = new StmtVisitor();
+      JimpleBodyConverterState state =
+          new JimpleBodyConverterState(
+              sourcePath, util, declaringClass, identifierFactory, unresolvedBranches, locals);
+      StmtVisitor stmtVisitor = new StmtVisitor(state);
       final JimpleParser.StatementsContext statements = method_body_contentsContext.statements();
       if (statements != null && statements.statement() != null) {
         List<Stmt> currentStmtList = new ArrayList<>();
@@ -284,457 +289,6 @@ public class LazyJimpleMethodSource implements BodySource {
       }
 
       return bodyBuilder.build();
-    }
-
-    /**
-     * Statement visitor - copied from JimpleConverter.MethodVisitor.StmtVisitor This needs access
-     * to the parser's state
-     */
-    private class StmtVisitor extends sootup.jimple.JimpleBaseVisitor<Stmt> {
-      final ValueVisitor valueVisitor = new ValueVisitor();
-
-      @Override
-      public Stmt visitStatement(JimpleParser.StatementContext ctx) {
-        final JimpleParser.StmtContext stmtCtx = ctx.stmt();
-        if (stmtCtx == null) {
-          throw new ResolveException(
-              "Couldn't parse Stmt.", sourcePath, JimpleConverterUtil.buildPositionFromCtx(ctx));
-        }
-        return visitStmt(stmtCtx);
-      }
-
-      @Override
-      @NonNull
-      public Stmt visitStmt(JimpleParser.StmtContext ctx) {
-        StmtPositionInfo pos = new SimpleStmtPositionInfo(ctx.start.getLine());
-
-        if (ctx.BREAKPOINT() != null) {
-          return Jimple.newBreakpointStmt(pos);
-        } else {
-          if (ctx.ENTERMONITOR() != null) {
-            return Jimple.newEnterMonitorStmt(valueVisitor.visitImmediate(ctx.immediate()), pos);
-          } else if (ctx.EXITMONITOR() != null) {
-            return Jimple.newExitMonitorStmt(valueVisitor.visitImmediate(ctx.immediate()), pos);
-          } else if (ctx.SWITCH() != null) {
-            return handleSwitchStmt(ctx, pos);
-          } else {
-            final JimpleParser.AssignmentsContext assignments = ctx.assignments();
-            if (assignments != null) {
-              return handleAssignment(assignments, pos);
-            } else if (ctx.IF() != null) {
-              return handleIfStmt(ctx, pos);
-            } else if (ctx.goto_stmt() != null) {
-              return handleGotoStmt(ctx, pos);
-            } else if (ctx.NOP() != null) {
-              return Jimple.newNopStmt(pos);
-            } else if (ctx.RET() != null) {
-              return Jimple.newRetStmt(valueVisitor.visitImmediate(ctx.immediate()), pos);
-            } else if (ctx.RETURN() != null) {
-              return handleReturnStmt(ctx, pos);
-            } else if (ctx.THROW() != null) {
-              return Jimple.newThrowStmt(valueVisitor.visitImmediate(ctx.immediate()), pos);
-            } else if (ctx.invoke_expr() != null) {
-              return Jimple.newInvokeStmt(
-                  (sootup.core.jimple.common.expr.AbstractInvokeExpr)
-                      valueVisitor.visitInvoke_expr(ctx.invoke_expr()),
-                  pos);
-            }
-          }
-        }
-        throw new ResolveException(
-            "Unknown Stmt.", sourcePath, JimpleConverterUtil.buildPositionFromCtx(ctx));
-      }
-
-      private Stmt handleSwitchStmt(JimpleParser.StmtContext ctx, StmtPositionInfo pos) {
-        Immediate key = valueVisitor.visitImmediate(ctx.immediate());
-        List<sootup.core.jimple.common.constant.IntConstant> lookup = new ArrayList<>();
-        List<String> targetLabels = new ArrayList<>();
-        int min = Integer.MAX_VALUE;
-        String defaultLabel = null;
-
-        for (JimpleParser.Case_stmtContext it : ctx.case_stmt()) {
-          final JimpleParser.Case_labelContext case_labelContext = it.case_label();
-          if (case_labelContext.getText() != null && case_labelContext.DEFAULT() != null) {
-            if (defaultLabel == null) {
-              defaultLabel = it.goto_stmt().label_name.getText();
-            } else {
-              throw new ResolveException(
-                  "Only one default label is allowed!",
-                  sourcePath,
-                  JimpleConverterUtil.buildPositionFromCtx(ctx));
-            }
-          } else if (case_labelContext.integer_constant().getText() != null) {
-            final int value = Integer.parseInt(case_labelContext.integer_constant().getText());
-            min = Math.min(min, value);
-            lookup.add(sootup.core.jimple.common.constant.IntConstant.getInstance(value));
-            targetLabels.add(it.goto_stmt().label_name.getText());
-          } else {
-            throw new ResolveException(
-                "Label is invalid.", sourcePath, JimpleConverterUtil.buildPositionFromCtx(ctx));
-          }
-        }
-        targetLabels.add(defaultLabel);
-
-        sootup.core.jimple.javabytecode.stmt.JSwitchStmt switchStmt;
-        if (ctx.SWITCH().getText().charAt(0) == 't') {
-          int high = min + lookup.size() - 1;
-          switchStmt = Jimple.newTableSwitchStmt(key, min, high, pos);
-        } else {
-          switchStmt = Jimple.newLookupSwitchStmt(key, lookup, pos);
-        }
-        unresolvedBranches.put(switchStmt, targetLabels);
-        return switchStmt;
-      }
-
-      private Stmt handleAssignment(
-          JimpleParser.AssignmentsContext assignments, StmtPositionInfo pos) {
-        if (assignments.COLON_EQUALS() != null) {
-          Local left = getLocal(assignments.local.getText());
-
-          sootup.core.jimple.common.ref.IdentityRef ref;
-          final JimpleParser.Identity_refContext identityRefCtx = assignments.identity_ref();
-          if (identityRefCtx.caught != null) {
-            ref = sootup.java.core.language.JavaJimple.newCaughtExceptionRef();
-          } else {
-            final String type = assignments.identity_ref().type().getText();
-            if (identityRefCtx.parameter_idx != null) {
-              int idx = Integer.parseInt(identityRefCtx.parameter_idx.getText());
-              ref = Jimple.newParameterRef(util.getType(type), idx);
-            } else {
-              if (declaringClass.toString().equals(type)) {
-                ref = Jimple.newThisRef(declaringClass);
-              } else {
-                ref = Jimple.newThisRef(util.getClassType(type));
-              }
-            }
-          }
-          return Jimple.newIdentityStmt(left, ref, pos);
-
-        } else if (assignments.EQUALS() != null) {
-          LValue left =
-              assignments.local != null
-                  ? getLocal(assignments.local.getText())
-                  : (LValue)
-                      valueVisitor.visitReference(assignments.reference());
-
-          final Value right = valueVisitor.visitValue(assignments.value());
-          return Jimple.newAssignStmt(left, right, pos);
-        } else {
-          throw new ResolveException(
-              "Invalid assignment.",
-              sourcePath,
-              JimpleConverterUtil.buildPositionFromCtx(methodContext));
-        }
-      }
-
-      private Stmt handleIfStmt(JimpleParser.StmtContext ctx, StmtPositionInfo pos) {
-        final BranchingStmt stmt =
-            Jimple.newIfStmt(
-                (sootup.core.jimple.common.expr.AbstractConditionExpr)
-                    valueVisitor.visitBool_expr(ctx.bool_expr()),
-                pos);
-        unresolvedBranches.put(
-            stmt, Collections.singletonList(ctx.goto_stmt().label_name.getText()));
-        return stmt;
-      }
-
-      private Stmt handleGotoStmt(JimpleParser.StmtContext ctx, StmtPositionInfo pos) {
-        final BranchingStmt stmt = Jimple.newGotoStmt(pos);
-        unresolvedBranches.put(
-            stmt, Collections.singletonList(ctx.goto_stmt().label_name.getText()));
-        return stmt;
-      }
-
-      private Stmt handleReturnStmt(JimpleParser.StmtContext ctx, StmtPositionInfo pos) {
-        if (ctx.immediate() == null) {
-          return Jimple.newReturnVoidStmt(pos);
-        } else {
-          return Jimple.newReturnStmt(valueVisitor.visitImmediate(ctx.immediate()), pos);
-        }
-      }
-
-      // ValueVisitor inner class needs to be copied here as well
-      // This is a simplified version - the full implementation from JimpleConverter would be needed
-      private class ValueVisitor
-          extends sootup.jimple.JimpleBaseVisitor<Value> {
-        // Full implementation would mirror JimpleConverter.MethodVisitor.StmtVisitor.ValueVisitor
-        // For brevity, including key methods - full copy would be needed in production
-
-        @Override
-        public Value visitValue(JimpleParser.ValueContext ctx) {
-          // Implementation copied from JimpleConverter
-          if (ctx.NEW() != null && ctx.base_type != null) {
-            final Type type = util.getType(ctx.base_type.getText());
-            if (!(type instanceof ReferenceType)) {
-              throw new ResolveException(
-                  type + " is not a ReferenceType.",
-                  sourcePath,
-                  JimpleConverterUtil.buildPositionFromCtx(ctx));
-            }
-            return Jimple.newNewExpr((ClassType) type);
-          } else if (ctx.NEWARRAY() != null) {
-            final Type type = util.getType(ctx.array_type.getText());
-            if (type instanceof VoidType || type instanceof NullType) {
-              throw new ResolveException(
-                  type + " can not be an ArrayType.",
-                  sourcePath,
-                  JimpleConverterUtil.buildPositionFromCtx(ctx));
-            }
-
-            Immediate dim =
-                visitImmediate(ctx.array_descriptor().immediate());
-            return sootup.java.core.language.JavaJimple.newNewArrayExpr(
-                type, dim, sootup.java.core.JavaIdentifierFactory.getInstance());
-          } else if (ctx.NEWMULTIARRAY() != null && ctx.immediate() != null) {
-            final Type type = util.getType(ctx.multiarray_type.getText());
-            if (!(type instanceof ReferenceType || type instanceof PrimitiveType)) {
-              throw new ResolveException(
-                  " Only base types are allowed",
-                  sourcePath,
-                  JimpleConverterUtil.buildPositionFromCtx(ctx));
-            }
-
-            List<Immediate> sizes =
-                ctx.immediate().stream()
-                    .map(this::visitImmediate)
-                    .collect(java.util.stream.Collectors.toList());
-            if (sizes.isEmpty()) {
-              throw new ResolveException(
-                  "The Size list must have at least one Element.",
-                  sourcePath,
-                  JimpleConverterUtil.buildPositionFromCtx(ctx));
-            }
-            ArrayType arrtype = view.getIdentifierFactory().getArrayType(type, sizes.size());
-            return Jimple.newNewMultiArrayExpr(arrtype, sizes);
-          } else if (ctx.nonvoid_cast != null && ctx.op != null) {
-            final Type type = util.getType(ctx.nonvoid_cast.getText());
-            Immediate val = visitImmediate(ctx.op);
-            return Jimple.newCastExpr(val, type);
-          } else if (ctx.INSTANCEOF() != null && ctx.op != null) {
-            final Type type = util.getType(ctx.nonvoid_type.getText());
-            Immediate val = visitImmediate(ctx.op);
-            return Jimple.newInstanceOfExpr(val, type);
-          }
-          return super.visitValue(ctx);
-        }
-
-        @Override
-        public Immediate visitImmediate(
-            JimpleParser.ImmediateContext ctx) {
-          if (ctx.identifier() != null) {
-            return getLocal(ctx.identifier().getText());
-          }
-          return visitConstant(ctx.constant());
-        }
-
-        @Override
-        public Value visitReference(JimpleParser.ReferenceContext ctx) {
-          if (ctx.array_descriptor() != null) {
-            Immediate idx =
-                visitImmediate(ctx.array_descriptor().immediate());
-            Local type = getLocal(ctx.identifier().getText());
-            return JavaJimple.newArrayRef(type, idx);
-          } else if (ctx.DOT() != null) {
-            String base = ctx.identifier().getText();
-            sootup.core.signatures.FieldSignature fs =
-                util.getFieldSignature(ctx.field_signature());
-            return Jimple.newInstanceFieldRef(getLocal(base), fs);
-          } else {
-            sootup.core.signatures.FieldSignature fs =
-                util.getFieldSignature(ctx.field_signature());
-            return Jimple.newStaticFieldRef(fs);
-          }
-        }
-
-        @Override
-        public Expr visitInvoke_expr(
-            JimpleParser.Invoke_exprContext ctx) {
-          List<Immediate> arglist = getArgList(ctx.arg_list(0));
-
-          if (ctx.nonstaticinvoke != null) {
-            Local base = getLocal(ctx.local_name.getText());
-            MethodSignature methodSig = util.getMethodSignature(ctx.method_signature(), ctx);
-
-            switch (ctx.nonstaticinvoke.getText().charAt(0)) {
-              case 'i':
-                return Jimple.newInterfaceInvokeExpr(base, methodSig, arglist);
-              case 'v':
-                return Jimple.newVirtualInvokeExpr(base, methodSig, arglist);
-              case 's':
-                return Jimple.newSpecialInvokeExpr(base, methodSig, arglist);
-              default:
-                throw new ResolveException(
-                    "Unknown Nonstatic Invoke.",
-                    sourcePath,
-                    JimpleConverterUtil.buildPositionFromCtx(ctx));
-            }
-          } else if (ctx.staticinvoke != null) {
-            MethodSignature methodSig = util.getMethodSignature(ctx.method_signature(), ctx);
-            return Jimple.newStaticInvokeExpr(methodSig, arglist);
-          } else if (ctx.dynamicinvoke != null) {
-            List<Type> bootstrapMethodRefParams = util.getTypeList(ctx.type_list());
-            MethodSignature bootstrapMethodRef =
-                view.getIdentifierFactory()
-                    .getMethodSignature(
-                        view.getIdentifierFactory()
-                            .getClassType(
-                                sootup.core.jimple.common.expr.JDynamicInvokeExpr
-                                    .INVOKEDYNAMIC_DUMMY_CLASS_NAME),
-                        ctx.STRING_CONSTANT().getText().replace("\"", ""),
-                        util.getType(ctx.name.getText()),
-                        bootstrapMethodRefParams);
-
-            MethodSignature methodRef = util.getMethodSignature(ctx.bsm, ctx);
-            List<Immediate> bootstrapArgs = getArgList(ctx.staticargs);
-
-            return Jimple.newDynamicInvokeExpr(
-                methodRef, bootstrapArgs, bootstrapMethodRef, getArgList(ctx.dyn_args));
-          }
-          throw new ResolveException(
-              "Malformed Invoke Expression.",
-              sourcePath,
-              JimpleConverterUtil.buildPositionFromCtx(ctx));
-        }
-
-        @Override
-        public sootup.core.jimple.common.constant.Constant visitConstant(
-            JimpleParser.ConstantContext ctx) {
-          // Full constant handling implementation from JimpleConverter would go here
-          // Abbreviated for space
-          if (ctx.integer_constant() != null) {
-            String intConst = ctx.integer_constant().getText();
-            int lastCharPos = intConst.length() - 1;
-            if (intConst.charAt(lastCharPos) == 'L' || intConst.charAt(lastCharPos) == 'l') {
-              intConst = intConst.substring(0, lastCharPos);
-              return sootup.core.jimple.common.constant.LongConstant.getInstance(
-                  Long.parseLong(intConst));
-            }
-            return sootup.core.jimple.common.constant.IntConstant.getInstance(
-                Integer.parseInt(intConst));
-          } else if (ctx.FLOAT_CONSTANT() != null) {
-            String floatStr = ctx.FLOAT_CONSTANT().getText();
-            int lastCharPos = floatStr.length() - 1;
-            if (floatStr.charAt(lastCharPos) == 'F' || floatStr.charAt(lastCharPos) == 'f') {
-              floatStr = floatStr.substring(0, lastCharPos);
-              return sootup.core.jimple.common.constant.FloatConstant.getInstance(
-                  Float.parseFloat(floatStr));
-            }
-
-            if (floatStr.charAt(0) == '#') {
-              switch (floatStr.substring(1)) {
-                case "Infinity":
-                  return sootup.core.jimple.common.constant.DoubleConstant.getInstance(
-                      Double.POSITIVE_INFINITY);
-                case "-Infinity":
-                  return sootup.core.jimple.common.constant.DoubleConstant.getInstance(
-                      Double.NEGATIVE_INFINITY);
-                case "NaN":
-                  return sootup.core.jimple.common.constant.DoubleConstant.getInstance(Double.NaN);
-              }
-            }
-
-            return sootup.core.jimple.common.constant.DoubleConstant.getInstance(
-                Double.parseDouble(floatStr));
-          } else if (ctx.CLASS() != null) {
-            final String text =
-                sootup.core.jimple.JimpleUtils.unescape(ctx.STRING_CONSTANT().getText());
-            return sootup.java.core.language.JavaJimple.newClassConstant(text);
-          } else if (ctx.STRING_CONSTANT() != null) {
-            final String text =
-                sootup.core.jimple.JimpleUtils.unescape(ctx.STRING_CONSTANT().getText());
-            return sootup.java.core.language.JavaJimple.newStringConstant(text);
-          } else if (ctx.BOOL_CONSTANT() != null) {
-            final char firstChar = ctx.BOOL_CONSTANT().getText().charAt(0);
-            return sootup.core.jimple.common.constant.BooleanConstant.getInstance(
-                firstChar == 't' || firstChar == 'T');
-          } else if (ctx.NULL() != null) {
-            return sootup.core.jimple.common.constant.NullConstant.getInstance();
-          }
-          // Additional constant types would be handled here
-          throw new ResolveException(
-              "Unknown Constant.", sourcePath, JimpleConverterUtil.buildPositionFromCtx(ctx));
-        }
-
-        @Override
-        public sootup.core.jimple.common.expr.AbstractBinopExpr visitBinop_expr(
-            JimpleParser.Binop_exprContext ctx) {
-          Immediate left = visitImmediate(ctx.left);
-          Immediate right = visitImmediate(ctx.right);
-
-          JimpleParser.BinopContext binopctx = ctx.binop();
-
-          if (binopctx.AND() != null) {
-            return new sootup.core.jimple.common.expr.JAndExpr(left, right);
-          } else if (binopctx.OR() != null) {
-            return new sootup.core.jimple.common.expr.JOrExpr(left, right);
-          } else if (binopctx.CMP() != null) {
-            return new sootup.core.jimple.common.expr.JCmpExpr(left, right);
-          } else if (binopctx.CMPG() != null) {
-            return new sootup.core.jimple.common.expr.JCmpgExpr(left, right);
-          } else if (binopctx.CMPL() != null) {
-            return new sootup.core.jimple.common.expr.JCmplExpr(left, right);
-          } else if (binopctx.CMPEQ() != null) {
-            return new sootup.core.jimple.common.expr.JEqExpr(left, right);
-          } else if (binopctx.CMPNE() != null) {
-            return new sootup.core.jimple.common.expr.JNeExpr(left, right);
-          } else if (binopctx.CMPGT() != null) {
-            return new sootup.core.jimple.common.expr.JGtExpr(left, right);
-          } else if (binopctx.CMPGE() != null) {
-            return new sootup.core.jimple.common.expr.JGeExpr(left, right);
-          } else if (binopctx.CMPLT() != null) {
-            return new sootup.core.jimple.common.expr.JLtExpr(left, right);
-          } else if (binopctx.CMPLE() != null) {
-            return new sootup.core.jimple.common.expr.JLeExpr(left, right);
-          } else if (binopctx.SHL() != null) {
-            return new sootup.core.jimple.common.expr.JShlExpr(left, right);
-          } else if (binopctx.SHR() != null) {
-            return new sootup.core.jimple.common.expr.JShrExpr(left, right);
-          } else if (binopctx.USHR() != null) {
-            return new sootup.core.jimple.common.expr.JUshrExpr(left, right);
-          } else if (binopctx.PLUS() != null) {
-            return new sootup.core.jimple.common.expr.JAddExpr(left, right);
-          } else if (binopctx.MINUS() != null) {
-            return new sootup.core.jimple.common.expr.JSubExpr(left, right);
-          } else if (binopctx.MULT() != null) {
-            return new sootup.core.jimple.common.expr.JMulExpr(left, right);
-          } else if (binopctx.DIV() != null) {
-            return new sootup.core.jimple.common.expr.JDivExpr(left, right);
-          } else if (binopctx.XOR() != null) {
-            return new sootup.core.jimple.common.expr.JXorExpr(left, right);
-          } else if (binopctx.MOD() != null) {
-            return new sootup.core.jimple.common.expr.JRemExpr(left, right);
-          }
-          throw new ResolveException(
-              "Unknown BinOp: " + binopctx.getText(),
-              sourcePath,
-              JimpleConverterUtil.buildPositionFromCtx(ctx));
-        }
-
-        @Override
-        public Expr visitUnop_expr(
-            JimpleParser.Unop_exprContext ctx) {
-          Immediate value = visitImmediate(ctx.immediate());
-          if (ctx.unop().NEG() != null) {
-            return Jimple.newNegExpr(value);
-          } else {
-            return Jimple.newLengthExpr(value);
-          }
-        }
-
-        @NonNull
-        private List<Immediate> getArgList(
-            JimpleParser.Arg_listContext ctx) {
-          if (ctx == null || ctx.immediate() == null) {
-            return Collections.emptyList();
-          }
-          final List<JimpleParser.ImmediateContext> immediates = ctx.immediate();
-          List<Immediate> arglist = new ArrayList<>(immediates.size());
-          for (JimpleParser.ImmediateContext immediate : immediates) {
-            arglist.add(visitImmediate(immediate));
-          }
-          return arglist;
-        }
-      }
     }
   }
 }
